Refactor Discord bot to use modular architecture with services and events
